Squashed commit of the following:
commit 7ba74304a5a04fa082dcc872889bbb9890e29a9e
Author: Madis <Madis0@users.noreply.github.com>
Date:   Sat Mar 11 13:09:32 2023 +0200

    rc2

commit 69570bf23b1eac0776469f4dcc3a62842c427e9f
Author: Madis <Madis0@users.noreply.github.com>
Date:   Fri Mar 10 14:24:11 2023 +0200

    3.6.0

commit 8a3597cf72654f717e0540e6a9ab5f3bda883051
Author: Madis <Madis0@users.noreply.github.com>
Date:   Thu Mar 2 13:26:12 2023 +0200

    3.6.0-beta.2

commit 526dfd046e93d82d4a85928650d0d0844b3a500a
Author: Madis <Madis0@users.noreply.github.com>
Date:   Fri Feb 24 16:32:44 2023 +0200

    3.6.0-beta.1

commit c2a4b8e375963024799ab46dd0b51dbca8c8eeff
Merge: 7d3df7a 10e2b29
Author: Madis <Madis0@users.noreply.github.com>
Date:   Fri Feb 24 16:09:22 2023 +0200

    Merge branch 'main' into 1.19.4

commit 7d3df7ad5d7aa99a287dd7bf841b02d9e4104b16
Merge: ee64809 5a2cc31
Author: Madis <Madis0@users.noreply.github.com>
Date:   Tue Feb 14 22:00:39 2023 +0200

    Merge branch 'main' into 1.19.4

commit ee64809b5d814148301943777f728d3259eb3706
Author: Madis <Madis0@users.noreply.github.com>
Date:   Mon Feb 13 19:42:49 2023 +0200

    23w06a

+++ src/main/java/io/github/madis0/PlayerProperties.java
@@ -19,7 +19,6 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Box;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.Difficulty;
-
 import java.text.DecimalFormat;
 import java.util.Comparator;
 import java.util.List;
@@ -304,14 +303,14 @@ public class PlayerProperties {
         levitationTime = levitationTimeRaw / (int) Calculations.getPrettyDivisor(maxLevitationTimeRaw, maxLevitationTime);
 
         belowBlockYRaw = playerEntity.getY();
-        double y = playerEntity.getY();
+        double y = (int)playerEntity.getY();
         double voidLimit = -128;
         BlockState state; //TODO: use material to determine fall damage
         if(playerEntity.getWorld() != null && (!playerEntity.isOnGround() || playerEntity.isSneaking())){
             var world = playerEntity.getWorld();
-            var x = playerEntity.getX();
-            var z = playerEntity.getZ();
-            while((state = world.getBlockState(new BlockPos(x, --y, z))).isAir() && y >= voidLimit) { //TODO: detect if block is solid and what height it has
+            var x = (int)playerEntity.getX();
+            var z = (int)playerEntity.getZ();
+            while((state = world.getBlockState(new BlockPos(x, (int)--y, z))).isAir() && y >= voidLimit) { //TODO: detect if block is solid and what height it has
                 belowBlockYRaw = y;
             }
         }
